+++ examples/react/next-routing/pages/index.tsx
@@ -33,10 +33,8 @@ type HitProps = {
 function Hit({ hit }: HitProps) {
   return (
     <>
-      <Link href="/other-page" passHref className="Hit-label">
-        <a>
-          <Highlight hit={hit} attribute="name" />
-        </a>
+      <Link href="/other-page" className="Hit-label">
+        <Highlight hit={hit} attribute="name" />
       </Link>
       <span className="Hit-price">${hit.price}</span>
     </>
@@ -66,6 +64,9 @@ export default function HomePage({ serverState, url }: HomePageProps) {
           router: createInstantSearchRouterNext({
             singletonRouter,
             serverUrl: url,
+            routerOptions: {
+              cleanUrlOnDispose: false,
+            },
           }),
         }}
         insights={true}