+++ packages/instantsearch.js/src/connectors/breadcrumb/connectBreadcrumb.ts
@@ -244,13 +244,16 @@ const connectBreadcrumb: BreadcrumbConnector = function connectBreadcrumb(
           hierarchicalFacetName
         );
 
-        return removeEmptyRefinementsFromUiState({
-          ...uiState,
-          hierarchicalMenu: {
-            ...uiState.hierarchicalMenu,
-            [hierarchicalFacetName]: path,
+        return removeEmptyRefinementsFromUiState(
+          {
+            ...uiState,
+            hierarchicalMenu: {
+              ...uiState.hierarchicalMenu,
+              [hierarchicalFacetName]: path,
+            },
           },
-        });
+          hierarchicalFacetName
+        );
       },
 
       getWidgetSearchParameters(searchParameters, { uiState }) {
@@ -333,29 +336,26 @@ function shiftItemsValues(array: BreadcrumbConnectorParamsItem[]) {
   }));
 }
 
-function removeEmptyRefinementsFromUiState(indexUiState: IndexUiState) {
-  const { hierarchicalMenu, ...indexUiStateBase } = indexUiState;
-
-  if (!hierarchicalMenu) {
+function removeEmptyRefinementsFromUiState(
+  indexUiState: IndexUiState,
+  attribute: string
+): IndexUiState {
+  if (!indexUiState.hierarchicalMenu) {
     return indexUiState;
   }
 
-  const connectorUiState = Object.keys(hierarchicalMenu).reduce(
-    (acc, key) => ({
-      ...acc,
-      ...(hierarchicalMenu[key].length > 0
-        ? { [key]: hierarchicalMenu[key] }
-        : {}),
-    }),
-    {}
-  );
+  if (
+    !indexUiState.hierarchicalMenu[attribute] ||
+    !indexUiState.hierarchicalMenu[attribute].length
+  ) {
+    delete indexUiState.hierarchicalMenu[attribute];
+  }
 
-  return {
-    ...indexUiStateBase,
-    ...(Object.keys(connectorUiState).length > 0
-      ? { hierarchicalMenu: connectorUiState }
-      : {}),
-  };
+  if (Object.keys(indexUiState.hierarchicalMenu).length === 0) {
+    delete indexUiState.hierarchicalMenu;
+  }
+
+  return indexUiState;
 }
 
 export default connectBreadcrumb;