+++ packages/react-instantsearch/src/ui/__tests__/HierarchicalMenu.test.tsx
@@ -80,7 +80,7 @@ describe('HierarchicalMenu', () => {
             class="ais-HierarchicalMenu-list"
           >
             <li
-              class="ais-HierarchicalMenu-item ais-HierarchicalMenu-item--parent ais-HierarchicalMenu-item--selected"
+              class="ais-HierarchicalMenu-item ais-HierarchicalMenu-item--selected ais-HierarchicalMenu-item--parent"
             >
               <a
                 class="ais-HierarchicalMenu-link ais-HierarchicalMenu-link--selected"
@@ -210,7 +210,7 @@ describe('HierarchicalMenu', () => {
               class="ais-HierarchicalMenu-list"
             >
               <li
-                class="ais-HierarchicalMenu-item ais-HierarchicalMenu-item--parent ais-HierarchicalMenu-item--selected"
+                class="ais-HierarchicalMenu-item ais-HierarchicalMenu-item--selected ais-HierarchicalMenu-item--parent"
               >
                 <a
                   class="ais-HierarchicalMenu-link ais-HierarchicalMenu-link--selected"
@@ -367,7 +367,7 @@ describe('HierarchicalMenu', () => {
             class="ais-HierarchicalMenu-list LIST"
           >
             <li
-              class="ais-HierarchicalMenu-item ITEM ais-HierarchicalMenu-item--parent PARENTITEM ais-HierarchicalMenu-item--selected SELECTEDITEM"
+              class="ais-HierarchicalMenu-item ITEM ais-HierarchicalMenu-item--selected SELECTEDITEM ais-HierarchicalMenu-item--parent PARENTITEM"
             >
               <a
                 class="ais-HierarchicalMenu-link LINK ais-HierarchicalMenu-link--selected SELECTEDITEMLINK"