+++ packages/vue-instantsearch/src/components/HierarchicalMenuList.vue
@@ -12,8 +12,8 @@
       :key="item.value"
       :class="[
         suit('item'),
-        item.data && item.data.length > 0 && suit('item', 'parent'),
         item.isRefined && suit('item', 'selected'),
+        item.data && item.data.length > 0 && suit('item', 'parent'),
       ]"
     >
       <a
@@ -24,7 +24,6 @@
         <span :class="suit('label')">{{ item.label }}</span>
         <span :class="suit('count')">{{ item.count }}</span>
       </a>
-
       <hierarchical-menu-list
         v-if="item.data"
         :items="item.data"