+++ packages/create-instantsearch-app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.3](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.3.2...create-instantsearch-app@7.3.3) (2023-12-19)
+
+**Note:** Version bump only for package create-instantsearch-app
+
+
+
+
+
+## [7.3.2](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.3.1...create-instantsearch-app@7.3.2) (2023-12-07)
+
+**Note:** Version bump only for package create-instantsearch-app
+
+
+
+
+
+## [7.3.1](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.3.0...create-instantsearch-app@7.3.1) (2023-11-28)
+
+**Note:** Version bump only for package create-instantsearch-app
+
+
+
+
+
 # [7.3.0](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.2.1...create-instantsearch-app@7.3.0) (2023-10-26)